Download a register

Downloads a whole register (or reads it from a file) and constructs an object that can be interrogated for its records, entries, items, schema, links to other registers, etc.

You should probably run rr_snapshot() on the output before using it.

rr_register(name = NULL, phase = c("beta", "alpha"), file = NULL,
  write = FALSE, dest_path = NULL, parse_datetimes = FALSE,
  quiet = TRUE)

Arguments

name
Character, the name of the register.

phase
Character, one of "beta" or "alpha"

file
Character, file path or URL, passed on to readr::read_lines() if name is not provided.

write
Logical, whether to write the RSF file to disk. If TRUE, either name or dest_path must be provided.

dest_path
Character, path and file name to write the RSF to.

parse_datetimes
Logical, whether to parse ISO8601 strings as datetimes with parsedate::parse_iso_8601(), otherwise leave as a string. Partial datetimes are parsed as the earliest possible datetime, e.g. "2018" becomes "2018-01-01 UTC".

quiet
Logical, if TRUE does not print messages to the console.

Value

An S3 object of class register

Examples

rr_register("country")

#> $root_hash
#> [1] "sha-256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
#> [2] "sha-256:725bf37d0ee609f7a9cb3e42af574443df9215c85772a1626eaa93cdd505e71f"
#> 
#> $entries
#> # A tibble: 219 x 5
#>    `entry-number` type   key                 timestamp           `hash-list`
#>             <int> <chr>  <chr>               <dttm>              <list>     
#>  1              1 system name                2017-07-17 10:59:47 <chr [1]>  
#>  2              2 system custodian           2017-07-17 10:59:47 <chr [1]>  
#>  3              3 system field:country       2017-01-10 17:16:07 <chr [1]>  
#>  4              4 system field:name          2017-01-10 17:16:07 <chr [1]>  
#>  5              5 system field:official-name 2017-01-10 17:16:07 <chr [1]>  
#>  6              6 system field:citizen-names 2017-01-10 17:16:07 <chr [1]>  
#>  7              7 system field:start-date    2017-01-10 17:16:07 <chr [1]>  
#>  8              8 system field:end-date      2017-01-10 17:16:07 <chr [1]>  
#>  9              9 system register:country    2016-08-04 14:45:41 <chr [1]>  
#> 10             10 system field:start-date    2017-08-29 11:30:00 <chr [1]>  
#> # ... with 209 more rows
#> 
#> $items
#> # A tibble: 219 x 3
#>    json       hash                                                data         
#>    <list>     <chr>                                               <list>       
#>  1 <list [4]> d97d6b34bc572e334cbd7898f785b72947557d9dbea5997707… <tibble [1 ×…
#>  2 <list [5]> 5c4728f439f6cbc6c7eea42992b858afc78c182962ba35d169… <tibble [1 ×…
#>  3 <list [4]> b00a1d254e6d187190abe43c8d87c76189d3e18a1c64ea88e3… <tibble [1 ×…
#>  4 <list [5]> af8130b639bb8a675234c313dd763781a7da4e13b7529583d3… <tibble [1 ×…
#>  5 <list [4]> dc1d12943ea264de937468b254286e5ebd8acd316e21bf6670… <tibble [1 ×…
#>  6 <list [4]> ee78f485f8ad1a169745366da4057c8fc0102e9461b5c1f0d1… <tibble [1 ×…
#>  7 <list [4]> 7c0ff61a59fd3997ba2c9fb12565cf00d75ba8eccdf863aa76… <tibble [1 ×…
#>  8 <list [4]> 466d194d5100532edd115e3f0035967b09bc7b7f5fc444166d… <tibble [1 ×…
#>  9 <list [4]> f2a77c13d773efac02a5d06864b2bcf73bb310aa54819b6f00… <tibble [1 ×…
#> 10 <list [4]> 28f3129e67f9090a61d044bf39411f5689c3eaa4249ea8324e… <tibble [1 ×…
#> # ... with 209 more rows
#> 
#> $schema
#> $schema$names
#> # A tibble: 1 x 6
#>   `entry-number` type   key   timestamp           hash                    name 
#>            <int> <chr>  <chr> <dttm>              <chr>                   <chr>
#> 1              1 system name  2017-07-17 10:59:47 d3d8e15fbd410e08bd8969… coun…
#> 
#> $schema$custodians
#> # A tibble: 2 x 6
#>   `entry-number` type   key       timestamp           hash            custodian
#>            <int> <chr>  <chr>     <dttm>              <chr>           <chr>    
#> 1              2 system custodian 2017-07-17 10:59:47 6bdb76b1cdec0f… Tony Wor…
#> 2             12 system custodian 2017-11-02 11:18:00 aa98858fc2a8a9… David de…
#> 
#> $schema$fields
#> # A tibble: 8 x 11
#>   `entry-number` type   key    timestamp           hash    field datatype phase
#>            <int> <chr>  <chr>  <dttm>              <chr>   <chr> <chr>    <chr>
#> 1              3 system field… 2017-01-10 17:16:07 a303d0… coun… string   beta 
#> 2              4 system field… 2017-01-10 17:16:07 a7a9f2… name  string   beta 
#> 3              5 system field… 2017-01-10 17:16:07 5c4728… offi… string   beta 
#> 4              6 system field… 2017-01-10 17:16:07 494f6f… citi… string   beta 
#> 5              7 system field… 2017-01-10 17:16:07 1cff4c… star… datetime beta 
#> 6              8 system field… 2017-01-10 17:16:07 a557fa… end-… datetime beta 
#> 7             10 system field… 2017-08-29 11:30:00 f09c43… star… datetime beta 
#> 8             11 system field… 2017-08-29 11:31:00 c5845b… end-… datetime beta 
#> # ... with 3 more variables: register <chr>, cardinality <chr>, text <chr>
#> 
#> 
#> $data
#> # A tibble: 207 x 11
#>    `entry-number` type  key   timestamp           hash           country name  
#>             <int> <chr> <chr> <dttm>              <chr>          <chr>   <chr> 
#>  1             13 user  SU    2016-04-05 13:23:05 e94c4a9ab00d9… SU      USSR  
#>  2             14 user  DE    2016-04-05 13:23:05 e03f97c280620… DE      West …
#>  3             15 user  DD    2016-04-05 13:23:05 e1357671d0da2… DD      East …
#>  4             16 user  YU    2016-04-05 13:23:05 a074752a77011… YU      Yugos…
#>  5             17 user  CS    2016-04-05 13:23:05 0031f311f8726… CS      Czech…
#>  6             18 user  GB    2016-04-05 13:23:05 6b18693874513… GB      Unite…
#>  7             19 user  AF    2016-04-05 13:23:05 6bf7f01f268fa… AF      Afgha…
#>  8             20 user  AL    2016-04-05 13:23:05 9d04a7e04ac92… AL      Alban…
#>  9             21 user  DZ    2016-04-05 13:23:05 3548cdf528e45… DZ      Alger…
#> 10             22 user  AD    2016-04-05 13:23:05 14fcb5099f0ef… AD      Andor…
#> # ... with 197 more rows, and 4 more variables: `official-name` <chr>,
#> #   `citizen-names` <chr>, `start-date` <chr>, `end-date` <chr>
#> 
#> attr(,"class")
#> [1] "register"

rr_register("country", "beta")

#> $root_hash
#> [1] "sha-256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
#> [2] "sha-256:725bf37d0ee609f7a9cb3e42af574443df9215c85772a1626eaa93cdd505e71f"
#> 
#> $entries
#> # A tibble: 219 x 5
#>    `entry-number` type   key                 timestamp           `hash-list`
#>             <int> <chr>  <chr>               <dttm>              <list>     
#>  1              1 system name                2017-07-17 10:59:47 <chr [1]>  
#>  2              2 system custodian           2017-07-17 10:59:47 <chr [1]>  
#>  3              3 system field:country       2017-01-10 17:16:07 <chr [1]>  
#>  4              4 system field:name          2017-01-10 17:16:07 <chr [1]>  
#>  5              5 system field:official-name 2017-01-10 17:16:07 <chr [1]>  
#>  6              6 system field:citizen-names 2017-01-10 17:16:07 <chr [1]>  
#>  7              7 system field:start-date    2017-01-10 17:16:07 <chr [1]>  
#>  8              8 system field:end-date      2017-01-10 17:16:07 <chr [1]>  
#>  9              9 system register:country    2016-08-04 14:45:41 <chr [1]>  
#> 10             10 system field:start-date    2017-08-29 11:30:00 <chr [1]>  
#> # ... with 209 more rows
#> 
#> $items
#> # A tibble: 219 x 3
#>    json       hash                                                data         
#>    <list>     <chr>                                               <list>       
#>  1 <list [4]> d97d6b34bc572e334cbd7898f785b72947557d9dbea5997707… <tibble [1 ×…
#>  2 <list [5]> 5c4728f439f6cbc6c7eea42992b858afc78c182962ba35d169… <tibble [1 ×…
#>  3 <list [4]> b00a1d254e6d187190abe43c8d87c76189d3e18a1c64ea88e3… <tibble [1 ×…
#>  4 <list [5]> af8130b639bb8a675234c313dd763781a7da4e13b7529583d3… <tibble [1 ×…
#>  5 <list [4]> dc1d12943ea264de937468b254286e5ebd8acd316e21bf6670… <tibble [1 ×…
#>  6 <list [4]> ee78f485f8ad1a169745366da4057c8fc0102e9461b5c1f0d1… <tibble [1 ×…
#>  7 <list [4]> 7c0ff61a59fd3997ba2c9fb12565cf00d75ba8eccdf863aa76… <tibble [1 ×…
#>  8 <list [4]> 466d194d5100532edd115e3f0035967b09bc7b7f5fc444166d… <tibble [1 ×…
#>  9 <list [4]> f2a77c13d773efac02a5d06864b2bcf73bb310aa54819b6f00… <tibble [1 ×…
#> 10 <list [4]> 28f3129e67f9090a61d044bf39411f5689c3eaa4249ea8324e… <tibble [1 ×…
#> # ... with 209 more rows
#> 
#> $schema
#> $schema$names
#> # A tibble: 1 x 6
#>   `entry-number` type   key   timestamp           hash                    name 
#>            <int> <chr>  <chr> <dttm>              <chr>                   <chr>
#> 1              1 system name  2017-07-17 10:59:47 d3d8e15fbd410e08bd8969… coun…
#> 
#> $schema$custodians
#> # A tibble: 2 x 6
#>   `entry-number` type   key       timestamp           hash            custodian
#>            <int> <chr>  <chr>     <dttm>              <chr>           <chr>    
#> 1              2 system custodian 2017-07-17 10:59:47 6bdb76b1cdec0f… Tony Wor…
#> 2             12 system custodian 2017-11-02 11:18:00 aa98858fc2a8a9… David de…
#> 
#> $schema$fields
#> # A tibble: 8 x 11
#>   `entry-number` type   key    timestamp           hash    field datatype phase
#>            <int> <chr>  <chr>  <dttm>              <chr>   <chr> <chr>    <chr>
#> 1              3 system field… 2017-01-10 17:16:07 a303d0… coun… string   beta 
#> 2              4 system field… 2017-01-10 17:16:07 a7a9f2… name  string   beta 
#> 3              5 system field… 2017-01-10 17:16:07 5c4728… offi… string   beta 
#> 4              6 system field… 2017-01-10 17:16:07 494f6f… citi… string   beta 
#> 5              7 system field… 2017-01-10 17:16:07 1cff4c… star… datetime beta 
#> 6              8 system field… 2017-01-10 17:16:07 a557fa… end-… datetime beta 
#> 7             10 system field… 2017-08-29 11:30:00 f09c43… star… datetime beta 
#> 8             11 system field… 2017-08-29 11:31:00 c5845b… end-… datetime beta 
#> # ... with 3 more variables: register <chr>, cardinality <chr>, text <chr>
#> 
#> 
#> $data
#> # A tibble: 207 x 11
#>    `entry-number` type  key   timestamp           hash           country name  
#>             <int> <chr> <chr> <dttm>              <chr>          <chr>   <chr> 
#>  1             13 user  SU    2016-04-05 13:23:05 e94c4a9ab00d9… SU      USSR  
#>  2             14 user  DE    2016-04-05 13:23:05 e03f97c280620… DE      West …
#>  3             15 user  DD    2016-04-05 13:23:05 e1357671d0da2… DD      East …
#>  4             16 user  YU    2016-04-05 13:23:05 a074752a77011… YU      Yugos…
#>  5             17 user  CS    2016-04-05 13:23:05 0031f311f8726… CS      Czech…
#>  6             18 user  GB    2016-04-05 13:23:05 6b18693874513… GB      Unite…
#>  7             19 user  AF    2016-04-05 13:23:05 6bf7f01f268fa… AF      Afgha…
#>  8             20 user  AL    2016-04-05 13:23:05 9d04a7e04ac92… AL      Alban…
#>  9             21 user  DZ    2016-04-05 13:23:05 3548cdf528e45… DZ      Alger…
#> 10             22 user  AD    2016-04-05 13:23:05 14fcb5099f0ef… AD      Andor…
#> # ... with 197 more rows, and 4 more variables: `official-name` <chr>,
#> #   `citizen-names` <chr>, `start-date` <chr>, `end-date` <chr>
#> 
#> attr(,"class")
#> [1] "register"

path <- tempfile()
download.file("https://country.register.gov.uk/download-rsf", path)
rr_register(file = path)

#> $root_hash
#> [1] "sha-256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
#> [2] "sha-256:725bf37d0ee609f7a9cb3e42af574443df9215c85772a1626eaa93cdd505e71f"
#> 
#> $entries
#> # A tibble: 219 x 5
#>    `entry-number` type   key                 timestamp           `hash-list`
#>             <int> <chr>  <chr>               <dttm>              <list>     
#>  1              1 system name                2017-07-17 10:59:47 <chr [1]>  
#>  2              2 system custodian           2017-07-17 10:59:47 <chr [1]>  
#>  3              3 system field:country       2017-01-10 17:16:07 <chr [1]>  
#>  4              4 system field:name          2017-01-10 17:16:07 <chr [1]>  
#>  5              5 system field:official-name 2017-01-10 17:16:07 <chr [1]>  
#>  6              6 system field:citizen-names 2017-01-10 17:16:07 <chr [1]>  
#>  7              7 system field:start-date    2017-01-10 17:16:07 <chr [1]>  
#>  8              8 system field:end-date      2017-01-10 17:16:07 <chr [1]>  
#>  9              9 system register:country    2016-08-04 14:45:41 <chr [1]>  
#> 10             10 system field:start-date    2017-08-29 11:30:00 <chr [1]>  
#> # ... with 209 more rows
#> 
#> $items
#> # A tibble: 219 x 3
#>    json       hash                                                data         
#>    <list>     <chr>                                               <list>       
#>  1 <list [4]> d97d6b34bc572e334cbd7898f785b72947557d9dbea5997707… <tibble [1 ×…
#>  2 <list [5]> 5c4728f439f6cbc6c7eea42992b858afc78c182962ba35d169… <tibble [1 ×…
#>  3 <list [4]> b00a1d254e6d187190abe43c8d87c76189d3e18a1c64ea88e3… <tibble [1 ×…
#>  4 <list [5]> af8130b639bb8a675234c313dd763781a7da4e13b7529583d3… <tibble [1 ×…
#>  5 <list [4]> dc1d12943ea264de937468b254286e5ebd8acd316e21bf6670… <tibble [1 ×…
#>  6 <list [4]> ee78f485f8ad1a169745366da4057c8fc0102e9461b5c1f0d1… <tibble [1 ×…
#>  7 <list [4]> 7c0ff61a59fd3997ba2c9fb12565cf00d75ba8eccdf863aa76… <tibble [1 ×…
#>  8 <list [4]> 466d194d5100532edd115e3f0035967b09bc7b7f5fc444166d… <tibble [1 ×…
#>  9 <list [4]> f2a77c13d773efac02a5d06864b2bcf73bb310aa54819b6f00… <tibble [1 ×…
#> 10 <list [4]> 28f3129e67f9090a61d044bf39411f5689c3eaa4249ea8324e… <tibble [1 ×…
#> # ... with 209 more rows
#> 
#> $schema
#> $schema$names
#> # A tibble: 1 x 6
#>   `entry-number` type   key   timestamp           hash                    name 
#>            <int> <chr>  <chr> <dttm>              <chr>                   <chr>
#> 1              1 system name  2017-07-17 10:59:47 d3d8e15fbd410e08bd8969… coun…
#> 
#> $schema$custodians
#> # A tibble: 2 x 6
#>   `entry-number` type   key       timestamp           hash            custodian
#>            <int> <chr>  <chr>     <dttm>              <chr>           <chr>    
#> 1              2 system custodian 2017-07-17 10:59:47 6bdb76b1cdec0f… Tony Wor…
#> 2             12 system custodian 2017-11-02 11:18:00 aa98858fc2a8a9… David de…
#> 
#> $schema$fields
#> # A tibble: 8 x 11
#>   `entry-number` type   key    timestamp           hash    field datatype phase
#>            <int> <chr>  <chr>  <dttm>              <chr>   <chr> <chr>    <chr>
#> 1              3 system field… 2017-01-10 17:16:07 a303d0… coun… string   beta 
#> 2              4 system field… 2017-01-10 17:16:07 a7a9f2… name  string   beta 
#> 3              5 system field… 2017-01-10 17:16:07 5c4728… offi… string   beta 
#> 4              6 system field… 2017-01-10 17:16:07 494f6f… citi… string   beta 
#> 5              7 system field… 2017-01-10 17:16:07 1cff4c… star… datetime beta 
#> 6              8 system field… 2017-01-10 17:16:07 a557fa… end-… datetime beta 
#> 7             10 system field… 2017-08-29 11:30:00 f09c43… star… datetime beta 
#> 8             11 system field… 2017-08-29 11:31:00 c5845b… end-… datetime beta 
#> # ... with 3 more variables: register <chr>, cardinality <chr>, text <chr>
#> 
#> 
#> $data
#> # A tibble: 207 x 11
#>    `entry-number` type  key   timestamp           hash           country name  
#>             <int> <chr> <chr> <dttm>              <chr>          <chr>   <chr> 
#>  1             13 user  SU    2016-04-05 13:23:05 e94c4a9ab00d9… SU      USSR  
#>  2             14 user  DE    2016-04-05 13:23:05 e03f97c280620… DE      West …
#>  3             15 user  DD    2016-04-05 13:23:05 e1357671d0da2… DD      East …
#>  4             16 user  YU    2016-04-05 13:23:05 a074752a77011… YU      Yugos…
#>  5             17 user  CS    2016-04-05 13:23:05 0031f311f8726… CS      Czech…
#>  6             18 user  GB    2016-04-05 13:23:05 6b18693874513… GB      Unite…
#>  7             19 user  AF    2016-04-05 13:23:05 6bf7f01f268fa… AF      Afgha…
#>  8             20 user  AL    2016-04-05 13:23:05 9d04a7e04ac92… AL      Alban…
#>  9             21 user  DZ    2016-04-05 13:23:05 3548cdf528e45… DZ      Alger…
#> 10             22 user  AD    2016-04-05 13:23:05 14fcb5099f0ef… AD      Andor…
#> # ... with 197 more rows, and 4 more variables: `official-name` <chr>,
#> #   `citizen-names` <chr>, `start-date` <chr>, `end-date` <chr>
#> 
#> attr(,"class")
#> [1] "register"

unlink(path)