Download all registers from an environment

Downloads all the registers listed in the register register of an environment ('alpha' or 'beta', meaning 'ready to use' or 'open for feedback').

Unlike rr_register() this function only downloads register, and does not read them from disk.

rr_registers(phase = c("beta", "alpha"), dir = NULL,
  parse_datetimes = FALSE, write = FALSE, dest_dir = phase,
  quiet = TRUE)

Arguments

phase
Character, one of "beta" or "alpha"

dir
Character, optional name of a directory of RSF files of registers. If NULL (default), the registers will be downloaded.

parse_datetimes
Logical, whether to parse ISO8601 strings as datetimes with parsedate::parse_iso_8601(), otherwise leave as a string. Partial datetimes are parsed as the earliest possible datetime, e.g. "2018" becomes "2018-01-01 UTC".

write
Logical, whether to write the RSF file to disk. If TRUE, either name or dest_path must be provided.

dest_dir
Character, name of a directory to save the RSF file of each register to. The file names will be the names of the registers.

quiet
Logical, if TRUE does not print messages to the console.

Value

An S3 object of class register

Examples

path <- fs::path(tempdir(), "beta")
names(rr_registers("beta", write = TRUE, dest_dir = path))

#>  [1] "country"                                            
#>  [2] "field"                                              
#>  [3] "territory"                                          
#>  [4] "government-organisation"                            
#>  [5] "government-service"                                 
#>  [6] "local-authority-sct"                                
#>  [7] "principal-local-authority"                          
#>  [8] "register"                                           
#>  [9] "datatype"                                           
#> [10] "internal-drainage-board"                            
#> [11] "registration-district"                              
#> [12] "statistical-geography-unitary-authority-wls"        
#> [13] "statistical-geography"                              
#> [14] "prison-estate"                                      
#> [15] "local-authority-nir"                                
#> [16] "statistical-geography-county-eng"                   
#> [17] "statistical-geography-london-borough-eng"           
#> [18] "statistical-geography-metropolitan-district-eng"    
#> [19] "statistical-geography-registration-district-eng"    
#> [20] "statistical-geography-registration-district-wls"    
#> [21] "statistical-geography-non-metropolitan-district-eng"
#> [22] "statistical-geography-unitary-authority-eng"        
#> [23] "allergen"                                           
#> [24] "statistical-geography-local-government-district-nir"
#> [25] "jobcentre"                                          
#> [26] "jobcentre-district"                                 
#> [27] "jobcentre-group"                                    
#> [28] "local-authority-type"                               
#> [29] "local-authority-eng"                                
#> [30] "qualification-assessment-method"                    
#> [31] "qualification-level"                                
#> [32] "qualification-sector-subject-area"                  
#> [33] "qualification-type"                                 
#> [34] "government-domain"                                  

names(rr_registers(dir = path))

#>  [1] "allergen"                                           
#>  [2] "country"                                            
#>  [3] "datatype"                                           
#>  [4] "field"                                              
#>  [5] "government-domain"                                  
#>  [6] "government-organisation"                            
#>  [7] "government-service"                                 
#>  [8] "internal-drainage-board"                            
#>  [9] "jobcentre-district"                                 
#> [10] "jobcentre-group"                                    
#> [11] "jobcentre"                                          
#> [12] "local-authority-eng"                                
#> [13] "local-authority-nir"                                
#> [14] "local-authority-sct"                                
#> [15] "local-authority-type"                               
#> [16] "principal-local-authority"                          
#> [17] "prison-estate"                                      
#> [18] "qualification-assessment-method"                    
#> [19] "qualification-level"                                
#> [20] "qualification-sector-subject-area"                  
#> [21] "qualification-type"                                 
#> [22] "register"                                           
#> [23] "registration-district"                              
#> [24] "statistical-geography-county-eng"                   
#> [25] "statistical-geography-local-government-district-nir"
#> [26] "statistical-geography-london-borough-eng"           
#> [27] "statistical-geography-metropolitan-district-eng"    
#> [28] "statistical-geography-non-metropolitan-district-eng"
#> [29] "statistical-geography-registration-district-eng"    
#> [30] "statistical-geography-registration-district-wls"    
#> [31] "statistical-geography-unitary-authority-eng"        
#> [32] "statistical-geography-unitary-authority-wls"        
#> [33] "statistical-geography"                              
#> [34] "territory"